ESBD: remove annoying text-selectability
Much of plot text (axes tick labels, trace labels, vertical markers)
could be selected by mouse, and it was getting annoying. So I set the
entire SVG plot to be unselectable.

Conversely the popup might want to be selected by user so I made it
selectable now.
(And since it's sibling of SVG, not child, no mouse event bubbling drama).