absorb Fader into ResponsivePlot.drawElements
It was a good run but the exiting class removal thing was not great when styles were applied to that class. So I had to introduce a data-join-exited attribute, which meant that the selectChildren had to be updated. This is the final encapsulation break which made me think that Fader was bad.

Since ResponsivePlot.drawElements was the only user of Fader, it made sense to absorb it!